Fix all models using updated_dt of flaw
Before this commit, it was assumed that the updated_dt of any model (affect, CVSS score, etc) was the same as the parent Flaw. This assumption is wrong, as it only holds true if Bugzilla sync is enabled. Otherwise, they are independent. One could modify an Affect in OSIDB without modifying the parent flaw, and the updated_dt of the Affect would be different than its flaw. This caused an error when trying to modify these models through OSIM in the UAT environment, which has the sync disabled, as we would always pass the flaw's older updated_dt in the PUT request, resulting in an error: "Save operation based on an outdated model instance".

With this commit, each model gets its own latest updated_dt instead of the flaw's before a POST or PUT request.

Closes OSIDB-2646.