Switch from summary to cve_description
This was already partially done as field from OSIDB called summary
is already shown as Description in OSIM. This is a final part to
completely stop using summary from OSIDB and start using new
dedicated field cve_description, which will be displayed only as a
Description on OSIM side.

Reasoning behind naming the field cve_description instead of just
description in OSIDB is that there is already a field description
in OSIDB (future comment_zero) andthis will help not to confuse between
those two fields.