Add empty workflow state in Flaws
In OSIDB's commit 59884a2839465db2b5e2ed3356a70b2dea2fbdd0 the empty
workflow state was added. This is not taken into account in OSIM, so
when trying to update a flaw that has an empty string as workflow state,
the type check will not even allow the call to OSIDB to be made.

This commit fixes this by adding the empty string as a possible workflow
state.

Closes OSIDB-2960.