Allow deleting CVSS scores
Right now it is not possible to delete a flaw's CVSS score through OSIM.
Even though there is a button to clear the field, when trying to save
the flaw it will show an error saying that this field cannot be null.

With this commit, when clearing the CVSS score field and saving the
flaw, it will delete the CVSS score through OSIDB's API, using the
DELETE method.

Closes OSIDB-1843.