Allow using default ordering in flaw list page
The flaw list page allows the user to order the results in descending or
ascending order by any of the fields in the header column. By default,
it orders the results by created date in descending order (newest flaws
first).

This commit adds the option to use the default ordering by OSIDB, by
toggling between descending, ascending, and no order when clicking on a
field in the header row. This is useful when using the search as OSIDB
will return the results ordered based on relevancy. For example, when
searching for a CVE, we may want to see the closest matches first,
instead of the most recent flaws.

Closes OSIDB-3187.