Adjust uniqueness check on ps_component/purl
In RedHatProductSecurity/osidb#1139 the uniqueness constraint for the
Affect model was changed. Instead of being unique on (flaw,
ps_update_stream, ps_component), now it is conditionally unique on
(flaw, ps_update_stream, ps_component) if no purl is present, or (flaw,
ps_update_stream, purl_normalized) if purl is present, where
purl_normalized is the purl with the qualifiers stripped.

This commit adjusts the uniqueness check for affects in OSIM as well,
taking those new constraints into consideration.